fix: move ManagerSaveException to tracker.exceptions and fix FileBackedTaskManagerTest history check
